+++ NEWS.md
@@ -4,7 +4,7 @@
 
 ### Bug fixes
 
- - fixes `fit_screen` (used in `stop_up` and `warn_up`) so that it does not mess with white spaces.
+ - fixes `fit_screen` so that it does not mess with white spaces.
 
 ### Name changes
 
@@ -22,6 +22,8 @@ try(check_value(x, "numeric scalar LE{z}"))
 #> Error: in check_value(x, "numeric scalar LE{z}"):
 #>  Value 'x' must be a numeric scalar lower than, or equal to, 32. Problem: it is strictly greater than 32.
 ```
+
+ - `stop_up` now accepts the argument msg which is an extra message displayed right after the error message.
 
 ## Version 1.2.3